+++ qt_storage_impl/include/qt_storage_impl/QtTaskStorageWriter.h
@@ -53,16 +53,16 @@ public:
 
 private:
     DBService& dbService;
-    long long addTaskQueryId{-1};
-    long long insertTagQueryId{-1};
-    long long removeTagQueryId{-1};
-    long long removeTaskQueryId{-1};
-    long long editQueryId{-1};
-    long long incrementSpentQueryId{-1};
-    long long decrementSpentQueryId{-1};
-    long long toggleCompletionQueryId{-1};
-    long long updatePrioritiesQueryId{-1};
-    long long editTagQueryId{-1};
+    qint64 addTaskQueryId{-1};
+    qint64 insertTagQueryId{-1};
+    qint64 removeTagQueryId{-1};
+    qint64 removeTaskQueryId{-1};
+    qint64 editQueryId{-1};
+    qint64 incrementSpentQueryId{-1};
+    qint64 decrementSpentQueryId{-1};
+    qint64 toggleCompletionQueryId{-1};
+    qint64 updatePrioritiesQueryId{-1};
+    qint64 editTagQueryId{-1};
 
     void insertTags(const QString& taskUuid, const std::list<Tag>& tagNames);
